Component refactoring benjamin (#13)
* rename according to conventions + styling

* minor change

* minor change

* style

* small test-changes

* change css style
