+++ src/components/App.js
@@ -1,32 +1,28 @@
 import "./App.css";
 import React from "react";
 import FormComponent from "./YAMLComponents/FormComponent";
-import Editor from "./mdPreviewComponents/editor";
-import {IndexPage} from "./frontPageComponents/indexPage";
-import {
-  BrowserRouter as Router,
-  Switch,
-  Route,
-} from "react-router-dom";
+
+import Editor from "./mdPreviewComponents/Editor";
+import { IndexPage } from "./frontPageComponents/IndexPage";
+import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 
 class App extends React.Component {
-
   render() {
     return (
-        /*To render the wanted component, add it to the div below:
-        * <FormComponent/>, <Editor/> or <IndexPage/>*/
+      /*To render the wanted component, add it to the div below:
+       * <FormComponent/>, <Editor/> or <IndexPage/>*/
       <Router>
-       <Switch>
+        <Switch>
           <Route exact path="/">
-            <div className="root container">
-              <IndexPage/>
-            </div>
+            <IndexPage />
           </Route>
           <Route path="/createNewLesson">
-            <FormComponent/> 
+            <FormComponent />
           </Route>
           <Route path="/editor">
-           <Editor/>
+            <div className="controlPanelPlacement">
+              <Editor />
+            </div>
           </Route>
         </Switch>
       </Router>
