+++ src/components/YAMLComponents/FormPage.js
@@ -1,7 +1,7 @@
 import React from "react";
 import YAMLformPage from "./YAMLformpage";
 import YMLformPage from "./YMLfilePage";
-import{Link} from "react-router-dom";
+import { Link } from "react-router-dom";
 
 const FormPage = props => {
   return (
@@ -24,14 +24,14 @@ const FormPage = props => {
         <div className="buttons">
           <button className="ui icon left labeled black button" type="button">
             <i aria-hidden="true" className="left arrow icon"></i>
-            <Link to="/">Tilbake</Link>  
+            <Link to="/">Tilbake</Link>
           </button>
           <button
             className="ui icon right labeled button toRight"
             type="button"
             onClick={props.mySubmitHandler}
           >
-           <Link to="/editor">Neste</Link> 
+            <Link to="/editor">Neste</Link>
             <i aria-hidden="true" className="right arrow icon"></i>
           </button>
         </div>
