+++ src/components/frontPageComponents/NavBar.js
