+++ src/components/mdPreviewComponents/Button.js
@@ -1,28 +1,16 @@
 import React from "react";
-import { Icon } from "semantic-ui-react";
-
 
 class Button extends React.Component {
-    onButtonClick = () => {
-        this.props.onButtonClick(this.props.output);
-
-        if (this.test === "") {
-            this.test = "inverted";
-        } else {
-            this.test = "";
-        }
-    };
-
-    render() {
-        return (
-            <div onClick={this.onButtonClick}>
-                <button>
-                    <i className={`${this.props.icon} icon`} />
-                    {this.props.title}
-                </button>
-            </div>
-        );
-    }
+  render() {
+    return (
+      <div onClick={() => this.props.onButtonClick(this.props.output)}>
+        <button>
+          {this.props.title}
+          <i className={`${this.props.icon} ${this.test} icon`} />
+        </button>
+      </div>
+    );
+  }
 }
 
 export default Button;
