+++ src/components/mdPreviewComponents/ControlPanel.js
@@ -1,52 +1,51 @@
 import React from "react";
-import Button from "./Button"
+import Button from "./Button";
 
 class ControlPanel extends React.Component {
+  handleButtonClick = value => {
+    this.props.handleButtonClick(value);
+  };
 
-    handleButtonClick = value => {
-        this.props.handleButtonClick(value);
-    };
-
-    render() {
-        return (
-            <div className="Buttons">
-                <div className="ui grid">
-                    <div className="column">
-                        <Button
-                            icon="bold"
-                            output=" **bold text here** "
-                            title=""
-                            onButtonClick={this.handleButtonClick}
-                        />
-                    </div>
-                    <div className="column">
-                        <Button
-                            icon="italic"
-                            output=" _italic text here_ "
-                            title=""
-                            onButtonClick={this.handleButtonClick}
-                        />
-                    </div>
-                    <div className="grid">
-                        <Button
-                            icon=""
-                            output=" {.activity} "
-                            title="Activity"
-                            onButtonClick={this.handleButtonClick}
-                        />
-                    </div>
-                    <div className="grid">
-                        <Button
-                            icon=""
-                            output=" {.intro} "
-                            title="Introduction"
-                            onButtonClick={this.handleButtonClick}
-                        />
-                    </div>
-                </div>
-            </div>
-        );
-    }
+  render() {
+    return (
+      <div className="Buttons">
+        <div className="ui segment four column grid">
+          <div className="column">
+            <Button
+              icon="bold"
+              output=" **bold text here** "
+              title=""
+              onButtonClick={this.handleButtonClick}
+            />
+          </div>
+          <div className="column">
+            <Button
+              icon="italic"
+              output=" _italic text here_ "
+              title=""
+              onButtonClick={this.handleButtonClick}
+            />
+          </div>
+          <div className="column">
+            <Button
+              icon=""
+              output=" {.activity} "
+              title="Activity"
+              onButtonClick={this.handleButtonClick}
+            />
+          </div>
+          <div className="column">
+            <Button
+              icon=""
+              output=" {.intro} "
+              title="Introduction"
+              onButtonClick={this.handleButtonClick}
+            />
+          </div>
+        </div>
+      </div>
+    );
+  }
 }
 
 export default ControlPanel;
