+++ src/index.css
@@ -11,40 +11,58 @@ body {
 
 code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
-  monospace;
+    monospace;
 }
 
-div {
+/* div {
   display: block;
-}
+} */
 
-.App {
+/* .App {
   display: flex;
   overflow-x: hidden;
   min-height: 100vh;
   background-color: #b1daae;
+} */
+
+/* CSS YAML Formpage */
+
+select {
+  cursor: pointer;
+}
+
+.ui.container {
+  margin-top: 5em;
 }
 
 /* CSS for Editor */
 
-.Editor {
+.test {
+  height: 40em;
+}
+
+.controlPanelPlacement {
+  margin-top: -8em;
+  width: auto;
+  height: auto;
+  background-color: #b1daae;
+  padding: 10em;
+}
+
+/* .Editor {
   display: flex;
   flex-direction: row;
   margin-left: auto;
   margin-right: auto;
   margin-top: 100px;
-}
+} */
 
 /* CSS for TextArea */
 
-.Buttons {
-  margin: 50px 0 -100px 100px !important;
-}
-
 .TextArea {
-  width: 600px;
-  height: 600px;
-  margin: 0 10px;
+  width: 100%;
+  height: 100%;
+  margin: auto;
   outline: none;
   resize: none;
   border-radius: 5px;
@@ -54,11 +72,14 @@ div {
 /* CSS for MDPreview */
 
 .MDPreview {
-  width: 600px;
-  height: 600px;
+  width: 100%;
+  height: 100%;
   border: 1px solid black;
-  margin: 0 10px;
+  border-radius: 5px;
+  margin: auto;
   background-color: white;
+  float: right;
+  padding: 15px;
 }
 
 /*This is the important css for the preview*/
